fix: improve result card reflow breakpoints (#3372)
#### Description of changes

Previously, the reflow behavior for cards (switching from "row headers to the left of content" to "row headers above content" at small widths) was driven using a media query over screen width. This worked OK in web and in reports, where cards are given the entire width of the screen to work with, but caused poor behavior in unified, where the cards are competing for screen width with the screenshot panel. In the most pathological width, you can see something like this:

**Before:**
![screenshot showing cards using only 3-4 characters per column for card row content](https://user-images.githubusercontent.com/376284/93527941-9d750280-f907-11ea-9f0f-78be5cfc3322.png)

This PR addresses this by redoing the reflow behavior for cards in terms of `display:flex` on the card rows, so the reflow behavior will be based on the width of the cards, not the width of the screen.

Per discussion with @ferBonnin, as part of this update, I've fixed up a few padding/margin sizes that were off from Figma's suggestions (in particular, we were using a width for the labels on android that was *way* larger than spec). With these changes, it was possible to combine the web/unified behavior and remove the unified-specific override by accepting web's padding being 4px higher than figma, so we did that (the result is still closer to the figma than the "before" case).

**I've reviewed each of these cases and encourage reviewer(s) to build the code locally and do the same:**
* Unified automated checks view
* Unified report exports
* Web fastpass automated checks
* Web fastpass needs review
* Web fastpass report exports
* Report package

#### Unified Automated Checks (100% zoom):
##### Before:
![unified-autochecks-before](https://user-images.githubusercontent.com/376284/93533155-563f3f80-f910-11ea-805b-0eb35f28d402.gif)

##### After:
![unified-autochecks-after](https://user-images.githubusercontent.com/376284/93533162-59d2c680-f910-11ea-965b-645228a994f5.gif)

#### Unified Report (150% zoom):
##### Before:
![unified-report-before](https://user-images.githubusercontent.com/376284/93533179-5fc8a780-f910-11ea-8726-52c2ebaa78a3.gif)

##### After:
![unified-report-after](https://user-images.githubusercontent.com/376284/93533187-622b0180-f910-11ea-81c2-13695a2a96cc.gif)

#### Web FastPass Automated Checks (125% zoom):
##### Before:
![web-autochecks-before](https://user-images.githubusercontent.com/376284/93533192-65be8880-f910-11ea-9adc-240335a5ac89.gif)

##### After:
![web-autochecks-after](https://user-images.githubusercontent.com/376284/93533203-68b97900-f910-11ea-9f39-7b0267d5c1db.gif)

#### Web FastPass Needs Review (120% zoom):
##### Before:
![web-needsreview-before](https://user-images.githubusercontent.com/376284/93533209-6ce59680-f910-11ea-9ba2-11fec6a22ce0.gif)

##### After:
![web-needsreview-after](https://user-images.githubusercontent.com/376284/93533212-6fe08700-f910-11ea-9284-748325159250.gif)


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: continuation from #3370
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS