Chore: Enable eslint rules (#3389)
* Chore: Enable eslint rules

- no-irregular-whitespace (0)
- no-var (7)
- no-regex-spaces (1)
- no-extra-boolean-cast (1)
- prefer-spread (3)

Also, fix null checks which appeared while addressing prefer-spread. This caused unit tests in assessment-store.test.ts to fail because actions in event-handler-list.ts were not invoked when the `eventArgs` parameter was undefined. The previous fix was to return from `invokeHandlers` if `eventArgs` was undefined, but the logic of the tests shows `eventArgs` is allowed to be undefined.

The proposed fix is to disallow the `sender` parameter of `invokeHandlers` from being undefined and to update the `FunctionPtr` interface to allow functions whose second (args) parameter may be undefined. This codifies the expectation that `sender` will always be present and that `eventArgs` is optional.