feat(android-left-nav): move unified command bar start over button (#3408)
When the leftNavBar feature flag is enabled, the Start over button appears to the right of the Export result button on the command bar. NVDA reads the components in the following order: Export result > Start over > Settings.

When the flag is disabled, the Start over button remains in its original position on the left side of the command bar. NVDA reads the components in the following order: Start over > Export result > Settings.