feat(android-left-nav): Remove generic key parameter from left nav types (#3411)
The genericity won't be necessary because it only adds complexity to have different stores and actions based on a key type. Instead, we will have one store that handles all the left nav actions and data, and only specific left nav items will be instantiated based on the need at the time.