feat(android-left-nav): Add content page info (#3427)
* feat(android-left-nav): Add content page info

- Add a type, `ContentPagesInfo`, which is a dictionary of `ContentPageInfo` objects referenced by `LeftNavItemKey`
- Create `ContentPanelDeps`, which contains a `ContentPagesInfo` instance
- Add the `description` property to the `TestConfig` object, which is a `JSX.Element` so that each page can have whatever style of content is required for its description.

`ContentPageInfo` contains the static information needed to render the content page displayed when a given left nav item is selected.

There are now two separate factories which create left nav dependency information based on the `TestConfig` object. The `TestConfig` object is designed so that differencial information about each test can be stored in one place. But we have elected not to simply pass the `TestConfig` objects around because it tends to bloat dependencies, and it makes determining what a component requires from the `TestConfig` less obvious.

* Use automated checks `HeaderSection` component as the `description` of the android automated checks content page