test(combined-report-ui): establish monorepo setup with new empty report-e2e-tests package (#3494)
#### Description of changes

This PR adds a new packages at `/packages/report-e2e-tests`. We'll be using this as part of the combined-report-ui feature to test report generation as it would work in real usage (ie, across package boundaries and with proper style embedding), but this also has the secondary benefit of setting up a lot of the infrastructure we'll be using for migrating the rest of the repo to the lerna/yarn workspaces monorepo format. Specifically, for generalized changes, it:

* Adds a new devDependency on lerna
* Establishes a `lerna.json` and a root level package.json `workspaces` property to establish the root of the repo as a yarn worktree/lerna monorepo (similar to the setup that service uses)
* Updates a variety of ignore/lint tool rules to ignore build output directories in subdirectories, not just the root
* Splits out the parts of `tsconfig.json` that will be shared among packages into `tsconfig.base.json` (the same layout that service uses)

Specific to the new `report-e2e-tests` package (empty with a placeholder test for now), it also:
* Creates the empty package, set up with typescript/jest/etc and a placeholder test
* Adds a root-level `yarn test:report:e2e` that runs the tests
* Adds a step to run the tests in CI/PR builds

Using yarn workspaces means that `yarn install` from the root like usual will cover all the requirements for installing the new package's dependencies. I used `lerna` to implement `yarn test:report:e2e` mostly as a POC (you could technically invoke `jest --projects packages/report-e2e-tests` and it would work equally well)

This PR **does not migrate /src to /packages/web/src** - instead, it leaves package.json to act as *both* the worktree root (the aggregator of workspaces) and *also* an individual workspace root representing web/unified/etc. This is supported by yarn, but not what we want to use long-term, since it's a bit confusing; this is an incremental step on the way towards using a layout closer to what service uses.

Out of scope for this PR:
* Migrating existing web code to `/packages/web`
* Migrating existing report package code to `/packages/report`
* Migrating the existing report integration tests to `/packages/report-e2e-tests`
* Extracting common Jest configuration to a `jest.config.base.js` like service does (this required a lot of changes, so I left it for a separate PR)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1783601
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS