feat(combined-report-ui): Introduce CollapsibleFailedResultsSection to contain Combined Results (#3545)
Currently, the CombinedReportSectionFactory does not provide a value for FailedInstancesSection. This PR introduces CollapsibleFailedResultsSection to act as the combined report's FailedInstancesSection. This is distinct from the existing FailedInstancesSection because it is collapsible and from the CollapsibleResultSection because it displays full result cards. A future PR will update the header to match the spec.