feat(combined-report-ui): Use URL count for combined report rule headers  (#3552)
Currently, the combined report's failed rule header counts are based on the number of cards displayed. This PR updates the combined report to use the sum of the number of URLs for the header count. To do this, OutcomeCounter is introduced and passed down to MinimalRuleHeader, which uses the counter to determine the number it displays. There should be no impact on non combined report header behavior.