chore(deps-dev): update build to node 14 (#3554)
#### Description of changes

Updates build to use node 14 + associated typings. 

Only required change was updating `mock-adb` to account for a behavior change where it's now an error to pass a number as content to `fs.writeFileSync` (it was doing this for a pidFile).

Note that this intentionally does not bump the engine requirements in package.json (that is, the project is cross-compatible with node 12 and 14). That's because service is currently blocked from updating to node 14, and I didn't want to introduce undue friction for working across the service and web repos. The package.json engines entries should be updated to enforce using 14 once service can update.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS