chore: remove use of Q from element-finder-by-position (#3616)
#### Description of changes

Web is only using Q (an obsolete implementation of Promises) in a couple of spots:

* element-finder-by-position
* frame-communicator
* a few analyzers in src/injected/analyzers/*

We'd like to remove these usages, so we can drop the obsolete Q dependency. This PR eliminates the first case (and also somewhat simplifies the implementation and significantly simplifies the tests).

The only thing that actually **uses** element-finder-by-position today is the injected `ScopingListener`, which uses it for the element picker of the scoping panel. I verified that this still works after the changes:

![gif showing the scoping picker working](https://user-images.githubusercontent.com/376284/98181843-471d4c80-1ed2-11eb-8570-8ac5625d2319.gif)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS