feat(android-needs-review): Filter GetCardSelectionViewData based on TestConfig (#3633)
Currently, GetCardSelectionViewData assumes that all results are applicable. This works in the web extension because we do not simultaneously hold failed and needs review results in the store. Now that we are adding needs review results to unified, that assumption is no longer valid, as we do not want to require the user to rescan their device when changing between Automated Checks and Needs Review.

To accommodate this scenario, this PR introduces the concept/type of a ResultsFilter and uses it to filter the selection data returned by GetCardSelectionViewData. This is done by adding a ResultsFilter to the appropriate unified TestConfig files. No filtering is required on the web side at this time.

The immediate user impact of this PR should be minimal--only the unified Needs Review tab's experience is impacted (Automated Check instances' highlights no longer show up).

Out of scope:

In the future, we may want to consider using the same ResultsFilter to filter CardsViewData as well
The unified Needs Review tab will still not display color contrast results