feat(android-needs-review): Fix Android Needs Review screenshot highlighting behavior (#3646)
PR #3633 introduced the concept of a ResultsFilter and used it to ensure that the proper elements are highlighted in Automated Checks and Needs Review when all cards are collapsed. Unfortunately, it failed to address the scenario where some cards are expanded or selected. This PR addresses those scenarios by ensuring that we always limit our resultsHighlightStatus data to reflect the currently selected tab's results.