fix: use axe-core tags to determine which WCAG links to show (#3709)
#### Description of changes

This PR changes the way we map from axe-core results/rules to WCAG guidance links.

Previously, we used a static mapping from rule ID to WCAG link. This mapping didn't have any guaranteed means of keeping in sync with axe-core's own WCAG tagging, so it drifts out of sync over time; #3703 notes several of the cases where we became inconsistent.

With this PR, we stop maintaining a rule->WCAG mapping and instead trust axe-core's tags to provide the rule->WCAG mapping; instead, we only maintain an axe tag -> WCAG link mapping.

The most meaningful implementation change is in converting `rule-to-links-mappings.ts` to `map-axe-tags-to-guidance-links.ts`; most the of the other changes are just updating the rest of the code to accomodate the change. Of note:

* `get-rule-inclusions` now looks for axe tags directly to determine if a rule should be excluded for being best-practice/experimental. There was one case (`form-field-multiple-labels`) where we were previously excluding a rule as "best practice" because its rule -> WCAG mapping said "best practice" even though axe-core didn't say so. I added an `explicitRuleOverrides` entry to make it more explicit why we exclude that case.
* `unique-landmark.ts` gained a `tags` entry, replacing the ad-hoc entry in the rules mapping
* The snapshot updates in the reporter E2Es are a good summary of the practical impact of the changes; mostly, it means that the WCAG links we present in the "passed"/"na" sections of reports change a bit.

Draft while we're going over the resulting changes in more detail

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: #3703
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS