Chore: Add eslint rule react/no-unused-state (#3849)
* Chore: Add eslint rule react/no-unused-state

This is one of the warnings found by lgtm.com.

While `isContextMenuVisible` was set, it's value was never read. Therefore, it could be removed.

There are two instances where the rule is disabled because it is finding false positives.

In launch-panel-header-click-handler.test.ts, I chose to leave `setStateMock` when it could also have been removed because this ensures that calling `openAdhocToolsPanel` has no unexpected side-effects related to the header object's state.

* Updating test snapshots