feat(focus-mvp-client): Add new test config for Tab Stops (#3872)
We need a new test config to add the Tab Stops test to AI-Android.

This PR adds such a config with the proper title and description. Because new components are not yet ready, it uses existing components for instancesSectionComponent and visualHelperSection. To avoid showing the new test, it introduces a tab stops feature flag as well.

This PR does not attempt to update how test configs are displayed to match the Tab Stops UI spec.