feat(focus-mvp-client): Add Tab Stops toggle e2e test (#3961)
Details
This PR adds an e2e test to validate that the tab stops visualization toggle sends enable/disable commands to the service.

Motivation
Feature spec

Context
Relies on the logging added in #3906. Some json filenames in the mock service were changed to match the endpoints called by the client.