refactor(side-panel): make closePreviewFeatures handling consistent with other side panels (#4046)
#### Details

This PR replaces the special handling for `onClosePreviewFeaturesPanel` with reuse of a common "close side panel" path already used to close the settings and scoping side panels. This was the only path that used `preview-features-actions`, so that file got removed as part of the PR.

##### Motivation

This finishes a collection of `refactor(side-panel)` PRs that @smoralesd did most of last year. I noticed it when trying to add `details-view-action-creator.ts` to the `strictNullChecks` list, where L88 hit an error due to this unfinished refactor; I decided that it seemed cleaner (net -83 LOC) to finish the refactor rather than update the types to work around the issue.

##### Context

Late followup to #2313, #2420, #2424

Verified manually that closing the preview features panel still appears to work as expected.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: contributes to #2869
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS