chore: Limit mac and windows unified e2e jobs to CI build (#4195)
Details
This PR limits the mac and windows unified e2e jobs to only run on the CI build, not the PR build.

Motivation
We're having some unified e2e test unreliability issues which will be investigated in an upcoming feature. In the meantime. this should help PR velocity without losing all of the benefits of our e2e tests.

Context
As discussed in the team retrospective.