feat(recommend-color): Able to recommend a color or two for insufficient color contrast ratio (#4360)
#### Details

This new feature will enable FastPass users to receive a color recommendation when running FastPass on webpages. The color recommendation will be displayed on the webpage when clicking a point of interest as well as in the FastPass output. 
When there is a color conflict, the current build informs users of this issue but does not recommend a way to fix the issue. With the new recommend color feature, a foreground and/or background is provided that will achieve success when testing for color contrast accessibility.

##### Motivation

MinT Project through Green River College

##### Context

![Screen Shot 2021-06-14 at 12 29 18 PM](https://user-images.githubusercontent.com/58874516/121969371-f5bb2280-cd28-11eb-85b2-6d49337b9814.png)
![Screen Shot 2021-06-14 at 12 29 28 PM](https://user-images.githubusercontent.com/58874516/121969374-f9e74000-cd28-11eb-99e1-19173deb733b.png)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS