chore(web-compliance-21h1): enable self-checking of FastPass needs review rules (#4378)
#### Details

This PR renames `window.A11YSelfValidator.validate()` to `window.selfFastPass.automatedChecks()` and adds a new `window.selfFastPass.needsReview()` which runs our FastPass Needs Review checks.

To do this, it rewrites the self validator to go through the `IssuesAdHocVisualization` and `NeedsReviewAdHocVisualization` config objects that contain the definitions of which rules should run for each case. This isn't totally accurate for Needs Review because it doesn't include the filtering performed by `UnifiedResultSender.filterNeedsReviewResults`; I decided to leave that out of scope for this PR, but it would be nice to do some sort of refactoring like we discussed in #4339 to enable reuse of that filtering logic. In the meantime, the main implication of not doing the filtering here is that some automated check failures would additionally show up as needs review failures, which isn't a big impact since we're generally automated-checks-clean.

##### Motivation

Wanted to run needs review checks as part of web compliance feature.

##### Context

I was initially concerned that the new dependencies on the visualizer configs would explode the size of the popup/insights bundles, but that doesn't seem to be the case in practice; it does increase them, but only by 1-2%.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1849910
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS