performance: switch web to es6 modules to improve build time and bundle size (#4376)
#### Details

This PR switches web's typescript configuration from outputting amd modules to outputting ES6 modules. This results in substantially smaller bundles and faster builds, since a huge amount of webpack boilerplate in the bundle output can be replaced by assuming that Chrome knows how to interpret an "import" statement.

Note that this is only updating typescript config, not webpack config. In particular, this means that the report package will still be generated as a umd package, compatible with node 12 where it's used by the service (through there is a slight change to codegen, hence the snapshot update).

##### Motivation

* Improves `yarn build` time by ~10% (from 70±2s to 63±2s)
* Improves dev bundle sizes by ~26% (injected.bundle.js from 21.6MB to 15.9MB)
* Improves prod bundle sizes by ~32% (injected.bundle.js from 2.88MB to 1.96MB)
* Required for later conversion to `swc-loader` in webpack (it doesn't support amd modules)

##### Context

I went through some smoke testing after these changes and didn't identify any behavioral changes (besides visibly improved perf with dev builds, especially around scan time).

The snapshot updates are due to slight changes in codegen. The `element-snapshot-formatter` update is because one of those codegen changes resulted in an office fabric component getting a different autogenerated id value; we normalized those already in id attributes, but this one was also referenced in a for attribute, so I updated the snapshot formatter to normalize those cases too.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS