feat(google-rules-mvp): call Android Service API based on atfaResults feature flag (#4397)
Details
Updated fetch-scan-results.ts to determine which Accessibility Insights for Android Service API to call, based on the status of the atfaResults feature flag:
* When the feature flag is enabled, it calls the v2 API.
* When the feature flag is disabled, it calls the original API (no version number).

Motivation
Partially addresses WI 1853273

Context
This is a follow-up to #4394, which handled reading the axe results from both v1 and v2 results formats. This PR updates the ScanResultsFetcher to call either of the two APIs in the Android Service, depending on whether the atfaResults feature flag has been enabled or disabled. The Android Service returns the proper data (v1 or v2) based on the API that is called.