add results_v2 json from scanning main screen of sunflower app (#4431)
Details
Added a results_v2.json for the mock-service to use. It was obtained from scanning the main page of the test Sunflower app.

Motivation
The Android mock-service uses a result.json file. Since the Accessibility Insights for Android Service is introducing a results_v2 endpoint that is called when the ATFA results feature flag is enabled, we need a similar results_v2.json that contains results in the modified format. This PR adds this file.

Context
E2E tests will be updated in another PR.