chore: remove dependency on q (#4447)
#### Details

`q` is a library that provides a standalone implementation of Promises. We've been gradually switching from it to native Promises for some time now. This PR finishes that work and removes the dependency on `q`.

Most of the interesting dependencies on Q were removed as part of #4155. The remaining interesting dependency was in `tab-stops-analyzer`, which used `Q.Promise`'s `progress()` callback in a way which isn't supported by native promises.

The only actual callback passed to the progress callback was the tab stops analyzer's own `onProgress` handler, which then calls out to `window.setTimeout` to implement some custom debouncing logic. This PR ripped out the `onProgress` implementation based on Q and replaced it with a more direct use of callbacks. While I was having to redo some of the test infrastructure anyway, I went ahead and replaced the custom debounce implementation with `lodash.debounce`; this required writing a `DebounceFaker` test helper to test with, but simplified the implementation substantially, which I think is a good tradeoff.

##### Motivation

Reduce unnecessary dependencies

##### Context

No user visible changes. I verified manually that paths through the tab-stops-analyzer (tab stops in adhoc, fastpass, assessment), rule-analyzer (individual assessment visualizers), and batched-analyzer (fastpass/assessment automated checks) seemed to have unchanged behavior.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS